Refactored SlotifyDataBase to DataAccessObject to better adhere to CA.
